fix textureId type (max textures is 512)

+++ LanguageBarrier/Game.cpp
@@ -325,7 +325,7 @@ FILE *clibFopenHook(const char *filename, const char *mode) {
 
 // TODO: I probably shouldn't be writing these in assembly given it looks like
 // they're all cdecl or thiscall anyway
-void gameLoadTexture(uint8_t textureId, void *buffer, size_t sz) {
+void gameLoadTexture(uint16_t textureId, void *buffer, size_t sz) {
   if (config["gamedef"]["gslPngLoadVersion"].get<std::string>() == "sghd") {
     ((SghdGslPngLoadProc)gameExeGslPngload)(textureId, buffer, sz);
   } else if (config["gamedef"]["gslPngLoadVersion"].get<std::string>() ==
